split up files to prep for meta overhaul
