+++ src/Components/Controls/ColorBrick/MoGuiColorBrick.cs
@@ -108,34 +108,4 @@ namespace MoGUI
             Brick.color = Value;
         }
     }
-
-    public class MoCaColor : MoGCArgs
-    {
-        public Func<Color> Value;
-        public MoCaColor(Color value,
-            //Func<object> text,
-            //Action onClickAction,
-            MoGuiMeta meta = null
-        ) : base(typeof(MoGuiColorBrick), meta)
-        {
-            Value = () => value;
-        }
-
-        public MoCaColor(Func<Color> value,
-            //Func<object> text,
-            //Action onClickAction,
-            MoGuiMeta meta = null
-        ) : base(typeof(MoGuiColorBrick), meta)
-        {
-            Value = value;
-        }
-
-    }
-
-    public class ColorBlockMeta : ControlMeta
-    {
-        public ColorBlockMeta(string name) : base(name) { MinSize(new Vector2(10, 10)); }
-
-
-    }
 }
