+++ src/Components/Controls/Input/MoGuiInput.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 using UnityEngine.EventSystems;
-using static TMPro.SpriteAssetUtilities.TexturePacker_JsonArray;
 
 namespace MoGUI
 {
@@ -322,58 +321,5 @@ namespace MoGUI
             }
                 
         }
-    }
-
-    public class MoCaInput : MoGCArgs
-    {
-        public object Value { get; set; }
-        public MoCaInput(Action<object> onEditAction,
-            Func<object> onUpdateAction,
-            Func<object> text,
-            string valType = "none",
-            MoGuiMeta meta = null
-        ) : base(typeof(MoGuiInput), meta:meta, text: text, onEditAction: onEditAction, onUpdateAction: onUpdateAction, valType: valType)
-        {
-
-        }
-
-        public MoCaInput(Action<object> onEditAction,
-            Func<object> onUpdateAction,
-            object text = null,
-            string valType = "none",
-            MoGuiMeta meta = null
-        ) : base(typeof(MoGuiInput), meta:meta, text: text, onEditAction: onEditAction, onUpdateAction: onUpdateAction, valType: valType)
-        {
-
-        }
-
-        public MoCaInput(object value,
-            
-            string valType = "none",
-            object text = null,
-            Action<object> onEditAction = null,
-            Func<object> onUpdateAction = null,
-            MoGuiMeta meta = null
-        ) : base(typeof(MoGuiInput), meta:meta, text: text, onEditAction: onEditAction, onUpdateAction: onUpdateAction, valType: valType)
-        {
-            Value = value;
-        }
-
-    }
-
-
-    public class InputMeta : ControlMeta
-    {
-
-        public Color background = GuiMeta.DefaultPanelColor.Shade;
-
-        public MoGuiColor textColor = GuiMeta.DefaultFontColor;
-
-        public InputMeta(string name) : base(name) { }
-
-
-
-
-
     }
 }
