+++ src/Components/MoGCArgs.cs
@@ -1,0 +1,74 @@
+﻿using System;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+using UnityEngine.EventSystems;
+
+
+namespace MoGUI
+{
+
+    public abstract class MoGCArgs
+    {
+        public Action OnClickAction;
+        public Func<object> OnUpdateAction;
+        public Action<object> OnEditAction;
+        public Func<object> Text;
+        public string ValType;
+        public string Orientation;
+        public string LabelPlacement;
+        public Type Type;
+        public MoGuiMeta Meta;
+
+
+        public MoGCArgs(Type type,
+            MoGuiMeta meta = null,
+            Func<object> value = null,
+             Action onClickAction = null,
+             Func<object> onUpdateAction = null,
+             Action<object> onEditAction = null,
+             Func<object> text = null,
+             string valType = "none"
+        )
+        {
+            Type = type;
+            Meta = meta;
+            OnClickAction = onClickAction;
+            OnUpdateAction = onUpdateAction;
+            OnEditAction = onEditAction;
+            Text = text;
+            ValType = valType;
+        }
+
+        public MoGCArgs(Type type, object value = null,
+             Action onClickAction = null,
+             Func<object> onUpdateAction = null,
+             Action<object> onEditAction = null,
+             object text = null,
+             string valType = null,
+            MoGuiMeta meta = null
+        )
+        {
+            Type = type;
+            Meta = meta;
+            OnClickAction = onClickAction;
+            OnUpdateAction = onUpdateAction;
+            OnEditAction = onEditAction;
+            Text = ConvertString(text);
+            ValType = valType;
+        }
+
+        public Func<object> ConvertString(object obj)
+        {
+            return () => obj;
+        }
+    }
+
+    public class ControlMeta : BlockMeta
+    {
+        public ControlMeta(string name) : base(name) { }
+
+
+    }
+
+}
