+++ src/Components/Resources/Layout/Layout.cs
@@ -7,166 +7,6 @@ using System.Xml.Linq;
 
 namespace MoGUI
 {
-    public class MoGuiRow : MoGuiLayoutBrick
-    {
-        public Dictionary<string, MoGuiCol> Columns;
-        private MoGuiPanel _parent;
-        public MoGuiRow(MoGuiPanel parent, string name, MoGuiMeta meta = null) : base(name)
-        {
-            _parent = parent;
-            if (meta == null)
-            {
-                Meta = _parent.Meta;
-            } else
-            {
-                Meta = meta;
-            }
-
-            Columns = new Dictionary<string, MoGuiCol>();
-            Obj = Create(name);
-            AddLayoutElement(Obj);
-            SetLayout();
-            SetParentRect(_parent.Container.GetComponent<RectTransform>());
-        }
-        protected override GameObject Create(string name)
-        {
-            GameObject layoutObject = new GameObject(Meta.PluginName + "_" + Meta.Name + "_" + "Row_" + name);
-            HorizontalLayoutGroup layoutGroup = layoutObject.AddComponent<HorizontalLayoutGroup>();
-            layoutGroup.padding = new RectOffset(Meta.TxtMargin, Meta.TxtMargin, Meta.TxtMargin, Meta.TxtMargin);
-            layoutGroup.spacing = Meta.TxtMargin;
-            layoutGroup.childForceExpandWidth = false;
-            layoutGroup.childForceExpandHeight = false;
-
-
-            if (MoGUIManager._LayoutDebug)
-            {
-                Image bg = layoutObject.AddComponent<Image>();
-                bg.color = MoGUIManager._LayoutDebugRowColor;
-            }
-            if (MoGui.TestMeta.Rows.background != null)
-            {
-                Image bg = layoutObject.AddComponent<Image>();
-                bg.color = (Color)MoGui.TestMeta.Rows.background;
-            }
-
-            return layoutObject;
-        }
-
-        public MoGuiCol AddColumn(string name)
-        {
-            MoGuiCol newCol = new MoGuiCol(this, name);
-            newCol.Obj.transform.SetParent(Obj.transform);
-            Columns.Add(name, newCol);
-            return newCol;
-        }
-
-        public MoGuiCol GetCol(string columnName)
-        {
-            MoGuiCol column;
-            if (Columns.ContainsKey(columnName))
-            {
-                column = Columns[columnName];
-            }
-            else
-            {
-                column = AddColumn(columnName);
-            }
-            return column;
-        }
-
-        public override void SetLayout()
-        {
-            minHeight = 10;
-            minWidth = 10;
-            flexibleHeight = 1;
-            flexibleWidth = 1;
-        }
-
-        public override void Update()
-        {
-            UpdateLayout();
-            foreach (var item in Columns)
-            {
-                item.Value.Update();
-            }
-        }
-    }
-
-    public class RowMeta : LayoutMeta
-    {
-
-        public RowMeta(string name) : base(name)
-        {
-            MinSize(GuiMeta.DefaultRowMinSize);
-            FlexSize(GuiMeta.DefaultRowFlex);
-        }
-
-    }
-
-    public class MoGuiCol : MoGuiLayoutBrick
-    {
-        private MoGuiRow _parent;
-        
-        public MoGuiCol(MoGuiRow parent, string name, MoGuiMeta meta = null) : base(name)
-        {
-            _parent = parent;
-            if (meta == null)
-            {
-                Meta = _parent.Meta;
-            }
-            else
-            {
-                Meta = meta;
-            }
-            Obj = Create(name);
-            AddLayoutElement(Obj);
-            SetLayout();
-            SetParentRect(_parent.Obj.GetComponent<RectTransform>());
-        }
-
-        protected override GameObject Create(string name)
-        {
-            GameObject layoutObject = new GameObject(Meta.PluginName + "_" + Meta.Name + "_" + _parent.Name + "_" + "Column_" + name);
-            VerticalLayoutGroup layoutGroup = layoutObject.AddComponent<VerticalLayoutGroup>();
-            layoutGroup.padding = new RectOffset(Meta.TxtMargin, Meta.TxtMargin, Meta.TxtMargin, Meta.TxtMargin);
-            layoutGroup.childForceExpandWidth = false;
-            layoutGroup.childForceExpandHeight = false;
-            layoutGroup.spacing = Meta.TxtMargin;
-
-            if (MoGUIManager._LayoutDebug)
-            {
-                Image bg = layoutObject.AddComponent<Image>();
-                bg.color = MoGUIManager._LayoutDebugColColor;
-            }
-
-            if (MoGui.TestMeta.Cols.background != null)
-            {
-                Image bg = layoutObject.AddComponent<Image>();
-                bg.color = (Color)MoGui.TestMeta.Cols.background;
-            }
-            return layoutObject;
-        }
-
-        public override void SetLayout()
-        {
-            minHeight = 10;
-            minWidth = 10;
-            flexibleWidth = 1;
-            flexibleHeight = 1;
-        }
-
-    }
-
-    public class ColMeta : LayoutMeta
-    {
-
-        public ColMeta(string name) : base(name)
-        {
-            MinSize(GuiMeta.DefaultColMinSize);
-            FlexSize(GuiMeta.DefaultColFlex);
-        }
-
-    }
 
     public abstract class MoGuiLayoutBrick
     {
